Dialer: Nearby places and reverse lookup improvements
* ContentProvider for cached places
* PlaceUtil: wrapper for interface methods, with automatical caching
* Place: add properties for address, email and website
* OsmApi: improve reverse lookup regex
* SearchChApi: add support for street and city

Nearby places:
* show distance in result list
* sort entries by distance
* move functionality to PlaceUtil (trying to make as little changes to AOSP code as possible)

Reverse lookup:
* Some carriers send the phone number without country code
  -> Convert the number in e164 format in SearchChApi
* change API method to getNamedPlaceByNumber(PhoneNumber phoneNumber)
* Call Log: support for reverse lookup
* Privacy: cache unsuccessful lookups as well

+++ src/com/android/dialer/omni/IPlacesAroundApi.java
@@ -25,16 +25,15 @@ public interface IPlacesAroundApi extends IRemoteApi {
 
     /**
      * Fetches and returns a list of named Places around the provided latitude and
-     * longitude parameters. The bounding box is calculated from lat-distance, lon-distance
-     * to lat+distance, lon+distance.
+     * longitude parameters.
      * This method is NOT asynchronous. Run it in a thread.
      *
      * @param name The name to search
      * @param lat Latitude of the point to search around
      * @param lon Longitude of the point to search around
-     * @param distance Max distance (polar coordinates)
+     * @param distance Max distance in meters
      * @return the list of matching places
      */
-    public List<Place> getNamedPlacesAround(String name, double lat, double lon, double distance);
+    public List<Place> getNamedPlacesAround(String name, double lat, double lon, long distance);
 
 }
